+++ tests/helper/daemon/mod.rs
@@ -1,9 +1,11 @@
 mod control;
 mod helper;
 mod setup;
+mod state;
 
 pub use control::*;
 pub use helper::*;
 pub use setup::*;
+pub use state::*;
 
 use super::*;