+++ tests/unix/mod.rs
@@ -1,5 +1,7 @@
 mod add;
+mod edit;
 mod kill;
+mod pause;
 /// Tests regarding state restoration from a previous run.
 mod restore;
 /// Tests for shutting down the daemon