+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 
+
+### Change
+
+- Make `GroupMessage` an enum to prevent impossible states.
+
 ## [0.17.2] - 2021-07-09
 
 ### Fix